mongodb error fix

+++ src/main/java/com/ewolff/microservice/customer/CustomerRepository.java
@@ -1,5 +1,6 @@
 package com.ewolff.microservice.customer;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -7,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "customer", path = "customer")
-public interface CustomerRepository extends MongoRepository<Customer, Long> {
+public interface CustomerRepository extends MongoRepository<Customer, BigInteger> {
 
 	List<Customer> findByName(@Param("name") String name);
